Create README.md
# Jamie-van-Cadsand

** GNU Declaration **

All programs and/or designs have been tested and approved by me. By using the program(s) or design(s) you agree to the standard user agreement. Which means that all or part of the downloads shown here are entirely at your own risk.
Any form of recourse is entirely or partially at your own risk.
Changes and/or adjustments, to both the software and hardware, which can cause damage are entirely for the person(s) who carried this out.  No liability can be claimed at all against the creator (Jamie van Cadsand). Everything is delivered As is. And is given entirely as a guideline or example. 

Jamie van Cadsand. The programmer.